Create README.md
Furcoder is an app that allows you to create "fur codes" that are used within the Furry Fandom to concisely tell a little about yourself. This originated in the Usenet days with the "Geek Code". You can also use it to decode a fur code you copied from websites and forums. Written  in C++ with the FOX toolkit. This is supplied with the distribution.